moved embeds into src
