+++ src/embeds/Bots/botCommands.js
