+++ src/embeds/Bots/botEmbeds.js
