+++ src/embeds/serverInfoEmbed.ts
